requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi
uvicorn
pydantic
python-dotenv

langchain
langchain-community
langchain-core
langchain-groq

faiss-cpu
sentence-transformers
pypdf
pymupdf
chromadb
groq
